[#5008] improve(core): reorganize the sql provider package to make it more clear (#5009)
This PR adjust and add some package for SQL provider to make the code
more separately and clear.

Make the code more clear.

Fix: #5008

No.

Exising tests.
